+++ CHANGELOG.md
@@ -12,6 +12,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [ 1.0.1 ] - [ xxxx-yy-zz ]
 
+### Added
+- Assembly declaration.
+
 ### Changed
 - Update to `mkdocs-material/9.6.5` and `mkdocstrings/0.28.2` versions in `requirements.txt`.
 - Do not hardcode `CMAKE_CXX_STANDARD`.